requirements.txt
aiosmtplib==2.0.2
annotated-types==0.7.0
anyio==4.13.0
APScheduler==3.11.2
asyncpg==0.31.0
bcrypt==4.0.1
blinker==1.9.0
certifi==2026.2.25
cffi==2.0.0
click==8.3.1
cryptography==46.0.5
dnspython==2.8.0
email_validator==2.2.0
fastapi==0.128.0
fastapi-mail==1.4.1
GeoAlchemy2==0.18.4
greenlet==3.3.2
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
Jinja2==3.1.6
MarkupSafe==3.0.3
packaging==26.0
passlib==1.7.4
psycopg2-binary==2.9.11
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.11.0
pydantic_core==2.41.5
PyJWT==2.10.1
python-dotenv==1.2.1
python-jose==3.5.0
python-multipart==0.0.20
PyYAML==6.0.3
redis==5.0.1
SQLAlchemy==2.0.46
SQLAlchemy-Utils==0.42.1
sqlmodel==0.0.37
starlette==0.50.0
typing_extensions==4.15.0
tzlocal==5.3.1
uvicorn==0.39.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==16.0